1. to_char as alias of to_string
2. delete TZM
