+++ src/query/expression/src/utils/date_helper.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::sync::LazyLock;
+
 use databend_common_exception::Result;
 use jiff::civil::date;
 use jiff::civil::datetime;
@@ -1315,43 +1317,47 @@ pub fn previous_or_next_day(dt: &Zoned, target: Weekday, is_previous: bool) -> i
     datetime_to_date_inner_number(dt) + dir * days_diff
 }
 
-pub struct PGDateTimeFormatter;
+static PG_STRFTIME_MAPPINGS: LazyLock<Vec<(&'static str, &'static str)>> = LazyLock::new(|| {
+    let mut mappings = vec![
+        ("YYYY", "%Y"),
+        ("YY", "%y"),
+        ("MMMM", "%B"),
+        ("MON", "%b"),
+        ("MM", "%m"),
+        ("DD", "%d"),
+        ("DY", "%a"),
+        ("HH24", "%H"),
+        ("HH12", "%I"),
+        ("AM", "%p"),
+        ("PM", "%p"),
+        ("MI", "%M"),
+        ("SS", "%S"),
+        ("FF", "%f"),
+        ("UUUU", "%G"),
+        ("TZHTZM", "%z"),
+        ("TZH:TZM", "%z"),
+        ("TZH", "%:::z"),
+    ];
+    // Sort by key length in descending order to ensure
+    // longer patterns are matched first and to avoid short patterns replacing part of long patterns prematurely.
+    mappings.sort_by(|a, b| b.0.len().cmp(&a.0.len()));
+    mappings
+});
 
-impl PGDateTimeFormatter {
-    pub fn format(dt: Zoned, format_string: &str) -> String {
-        let mut result = format_string.to_string();
-        let mut format_map: Vec<(&str, fn(&Zoned) -> String)> = Vec::new();
+#[inline]
+pub fn pg_format_to_strftime(pg_format_string: &str) -> String {
+    let mut result = pg_format_string.to_string();
+    for (pg_key, strftime_code) in PG_STRFTIME_MAPPINGS.iter() {
+        let pattern = if *pg_key == "MON" {
+            // should keep "month". Only "MON" as a single string escape it.
+            format!(r"(?i)\b{}\b", regex::escape(pg_key))
+        } else {
+            format!(r"(?i){}", regex::escape(pg_key))
+        };
+        let reg = regex::Regex::new(&pattern).expect("Failed to compile regex for format key");
 
-        format_map.push(("YYYY", |dt| dt.strftime("%Y").to_string()));
-        format_map.push(("YY", |dt| dt.strftime("%y").to_string()));
-        format_map.push(("MMMM", |dt| dt.strftime("%B").to_string()));
-        format_map.push(("MON", |dt| dt.strftime("%b").to_string()));
-        format_map.push(("MM", |dt| dt.strftime("%m").to_string()));
-        format_map.push(("DD", |dt| dt.strftime("%d").to_string()));
-        format_map.push(("DY", |dt| dt.strftime("%a").to_string()));
-        format_map.push(("HH24", |dt| dt.strftime("%H").to_string()));
-        format_map.push(("HH12", |dt| dt.strftime("%I").to_string()));
-        format_map.push(("AM", |dt| dt.strftime("%p").to_string()));
-        format_map.push(("PM", |dt| dt.strftime("%p").to_string()));
-        format_map.push(("MI", |dt| dt.strftime("%M").to_string()));
-        format_map.push(("SS", |dt| dt.strftime("%S").to_string()));
-        format_map.push(("FF", |dt| dt.strftime("%f").to_string()));
-        format_map.push(("TZH", |dt| {
-            dt.strftime("%z").to_string().chars().take(3).collect()
-        }));
-        format_map.push(("TZM", |dt| {
-            dt.strftime("%z")
-                .to_string()
-                .chars()
-                .skip(3)
-                .take(2)
-                .collect()
-        }));
-        format_map.push(("UUUU", |dt| dt.strftime("%G").to_string()));
-        for (key, func) in &format_map {
-            let reg = regex::Regex::new(&format!(r"(?i){}", key)).unwrap();
-            result = reg.replace_all(&result, func(&dt)).to_string();
-        }
-        result
+        // Use replace_all to substitute all occurrences of the PG key with the strftime code.
+        result = reg.replace_all(&result, *strftime_code).to_string();
     }
+    result
 }
