+++ src/query/functions/src/scalars/timestamp/Cargo.toml
@@ -11,4 +11,3 @@ databend-common-expression = { workspace = true }
 dtparse = { workspace = true }
 jiff = { workspace = true }
 num-traits = { workspace = true }
-regex = { workspace = true }
