+++ tests/sqllogictests/suites/query/functions/02_0012_function_datetimes.test
@@ -1471,17 +1471,17 @@ select date_format('2022-02-04T03:58:59', '%x'), strftime('2022-02-04T03:58:59',
 2022-02-04 03:58:59 2022-02-04 03:58:59
 
 query TTT
-select to_char('2024-04-05'::DATE, 'mon dd, yyyy');
+settings (date_format_style='Oracle') select to_char('2024-04-05'::DATE, 'mon dd, yyyy');
 ----
 Apr 05, 2024
 
 query TTT
-settings (timezone = 'Asia/Shanghai') select to_char('2024-04-05'::DATE, 'mon dd, yyyy TZH');
+settings (timezone = 'Asia/Shanghai', date_format_style='Oracle') select to_char('2024-04-05'::DATE, 'mon dd, yyyy TZH');
 ----
 Apr 05, 2024 +08
 
 query TTT
-select to_char('2024-04-05T00:00:00'::TIMESTAMP, 'mon dd HH12AM:MI:SS, yy');
+settings (date_format_style='Oracle') select to_char('2024-04-05T00:00:00'::TIMESTAMP, 'mon dd HH12AM:MI:SS, yy');
 ----
 Apr 05 12AM:00:00, 24
